+++ packages/common/src/create-payout-tx.ts
@@ -1,9 +1,15 @@
-import type { ChainedTxResult, TxResult, SpendableCoin, PayoutRule, Output, UTXO, TokenId, Fraction } from './types.js';
+import type {
+  ChainedTxResult, TxResult, SpendableCoin, PayoutRule, Output, UTXO, TokenId, Fraction,
+  InputParamsWithUnlocker,
+} from './types.js';
 import { InvalidProgramState, ValueError } from './exceptions.js';
 import { PayoutAmountRuleType, SpendableCoinType } from './constants.js';
-import { convertTokenIdToUint8Array } from './util.js';
+import {
+  inputParamsWithUnlockerToLibauthInputTemplate, outputToLibauthOutput,
+  calcAvailablePayoutFromIO,
+} from './util.js';
 import * as payoutBuilder from './payout-builder.js';
-import { calcAvailablePayoutFromLASourceOutputsAndOutputs, convertSpendableCoinsToLAInputsWithSourceOutput } from './util-libauth-dependent.js';
+import { spendableCoinToInputWithUnlocker } from './util-libauth-dependent.js';
 import * as libauth from './libauth.js';
 
 /**
@@ -15,11 +21,6 @@ export type CreatePayoutTxContext = {
   txfee_per_byte: Fraction | bigint;
   getOutputMinAmount (output: Output): bigint;
   getPreferredTokenOutputBCHAmount (output: Output): bigint | null;
-};
-
-type LibauthTemplateInputAndSourceOutput = {
-  input: libauth.InputTemplate<libauth.CompilerBCH>;
-  source_output: libauth.Output;
 };
 
 const convertToJSONSerializable = (v: any): any => {
@@ -60,41 +61,60 @@ export function createPayoutChainedTx (context: CreatePayoutTxContext, input_coi
   if (input_coins.length == 0) {
     throw new ValueError('number of input_coins should be greater than zero');
   }
-  const change_payout_rule = payout_rules.find((a) => a.type == PayoutAmountRuleType.CHANGE && a.spending_parameters != null);
-  if (change_payout_rule == null) {
-    throw new ValueError(`One change payout rule with spending parameters is needed to create-payout-chained-tx`);
-  }
-  if (change_payout_rule.spending_parameters?.type != SpendableCoinType.P2PKH) {
-    throw new ValueError(`The provided change payout rule should have its spending_parameters with P2PKH type.`);
+  const should_reuse_change_payouts = payout_rules.length !== 1;
+  let change_payout_rule;
+  if (should_reuse_change_payouts) {
+    change_payout_rule = payout_rules.find((a) => a.type == PayoutAmountRuleType.CHANGE && a.spending_parameters != null);
+    if (change_payout_rule == null) {
+      throw new ValueError(`One change payout rule with spending parameters is needed to create-payout-chained-tx`);
+    }
+    // more than a change address exists, change_payout_rule.spending_parameters is required
+    if (change_payout_rule.spending_parameters?.type != SpendableCoinType.P2PKH) {
+      throw new ValueError(`The provided change payout rule should have its spending_parameters with P2PKH type.`);
+    }
+  } else {
+    change_payout_rule = payout_rules.find((a) => a.type == PayoutAmountRuleType.CHANGE);
+    if (change_payout_rule == null) {
+      throw new ValueError(`One change payout rule is needed to create-payout-chained-tx`);
+    }
   }
   const result: ChainedTxResult = {
     chain: [],
     txfee: 0n,
     payouts: [],
   };
-  let input_items: LibauthTemplateInputAndSourceOutput[] = convertSpendableCoinsToLAInputsWithSourceOutput(input_coins);
+  const txparams: { locktime: number, version: number } = { locktime: 0, version: 2 };
+  let inputs: InputParamsWithUnlocker[] = input_coins.map((coin) => spendableCoinToInputWithUnlocker(coin, { sequence_number: 0 }));
   while (true) {
-    const { tx_result, done, unused_input_items } = createPayoutChainedTxSub(context, input_items, change_payout_rule, payout_rules);
+    const { tx_result, done, unused_inputs } = createPayoutChainedTxSub(context, inputs, change_payout_rule, payout_rules, txparams);
     result.txfee += tx_result.txfee;
     result.chain.push(tx_result);
     if (done) {
-      result.payouts = tx_result.payouts;
+      result.payouts = [ ...result.payouts, ...tx_result.payouts ];
       break;
     }
-    if (unused_input_items.length + tx_result.payouts.length >= input_items.length) {
-      /* c8 ignore next */
-      throw new InvalidProgramState(`unused_input_items.length + tx_result.payouts.length >= input_items.length`);
+    if (should_reuse_change_payouts) {
+      if (unused_inputs.length + tx_result.payouts.length >= inputs.length) {
+        /* c8 ignore next */
+        throw new InvalidProgramState(`unused_inputs.length + tx_result.payouts.length >= inputs.length`);
+      }
+      inputs = [ ...unused_inputs, ...tx_result.payouts.map((utxo) => spendableCoinToInputWithUnlocker({
+        type: SpendableCoinType.P2PKH,
+        output: utxo.output,
+        outpoint: utxo.outpoint,
+        key: (change_payout_rule as any).spending_parameters.key,
+      }, { sequence_number: 0 })) ];
+    } else {
+      if (unused_inputs.length >= inputs.length) {
+        /* c8 ignore next */
+        throw new InvalidProgramState(`unused_inputs.length >= inputs.length`);
+      }
+      inputs = unused_inputs;
+      result.payouts = [ ...result.payouts, ...tx_result.payouts ];
     }
-    input_items = [ ...unused_input_items, ...convertSpendableCoinsToLAInputsWithSourceOutput(tx_result.payouts.map((utxo) => ({
-      type: SpendableCoinType.P2PKH,
-      output: utxo.output,
-      outpoint: utxo.outpoint,
-      key: (change_payout_rule as any).spending_parameters.key,
-    }))) ];
   }
   return result;
 }
-
 
 /**
  * Create a transaction to payout a set of addresses/locking_bytecodes. Input coins provide the funding needed to build the payouts.
@@ -114,8 +134,9 @@ export function createPayoutTx (context: CreatePayoutTxContext, input_coins: Spe
   if (change_payout_rule == null) {
     throw new ValueError(`One change payout rule is needed to create-payout-tx`);
   }
-  let input_items: LibauthTemplateInputAndSourceOutput[] = convertSpendableCoinsToLAInputsWithSourceOutput(input_coins);
-  const { tx_result, done } = createPayoutChainedTxSub(context, input_items, change_payout_rule, payout_rules);
+  const inputs: InputParamsWithUnlocker[] = input_coins.map((coin) => spendableCoinToInputWithUnlocker(coin, { sequence_number: 0 }));
+  const txparams: { locktime: number, version: number } = { locktime: 0, version: 2 };
+  const { tx_result, done } = createPayoutChainedTxSub(context, inputs, change_payout_rule, payout_rules, txparams);
   if (!done) {
     throw new ValueError(`Too many inputs, Tx size limit reached, Cannot generate a single transaction to perform the payout.` );
   }
@@ -125,45 +146,32 @@ export function createPayoutTx (context: CreatePayoutTxContext, input_coins: Spe
 
 const utxoFromPayoutResult = (txhash: Uint8Array, a: { output: Output, output_index: number }): UTXO => ({ outpoint: { txhash, index: a.output_index }, output: a.output });
 
-const createPayoutChainedTxSub = (context: CreatePayoutTxContext, input_items: LibauthTemplateInputAndSourceOutput[], change_payout_rule: PayoutRule, payout_rules: PayoutRule[]): { tx_result: TxResult, done: boolean, unused_input_items: Array<{ input: libauth.InputTemplate<libauth.CompilerBCH>, source_output: libauth.Output }>  } => {
+const createPayoutChainedTxSub = (context: CreatePayoutTxContext, inputs: InputParamsWithUnlocker[], change_payout_rule: PayoutRule, payout_rules: PayoutRule[], txparams: { locktime: number, version: number }): { tx_result: TxResult, done: boolean, unused_inputs: InputParamsWithUnlocker[] } => {
   const txfee_per_byte = typeof context.txfee_per_byte == 'bigint' ?
     { numerator: context.txfee_per_byte, denominator: 1n } : context.txfee_per_byte as Fraction;
   if (!(txfee_per_byte.numerator >= 0n && txfee_per_byte.denominator > 0n)) {
     throw new ValueError('txfee should be greater than or equal to zero, txfee_per_byte type: Fraction | bigint');
   }
-  const makeLAOutputs = (outputs: Output[]): libauth.Output[] => {
-    return outputs.map((a) => ({
-      lockingBytecode: a.locking_bytecode,
-      token: a.token ? {
-        amount: a.token.amount < 0n ? context.getOutputMinAmount(a) : a.token.amount,
-        category: convertTokenIdToUint8Array(a.token.token_id),
-        nft: a.token.nft ? {
-          capability: a.token.nft.capability,
-          commitment: a.token.nft.commitment,
-        } : undefined,
-      } : undefined,
-      valueSatoshis: a.amount,
-    }))
-  };
   const max_tx_size: bigint = 100000n;
-  let unused_input_items: LibauthTemplateInputAndSourceOutput[] = input_items;
-  let sub_input_items: LibauthTemplateInputAndSourceOutput[] = [];
+  let unused_inputs: InputParamsWithUnlocker[] = inputs;
+  let sub_inputs: InputParamsWithUnlocker[] = [];
   let best_result = null;
   let best_result_done = false;
   let next_attempt_addtional_input_count = 450;
   let offset = 0;
   while (next_attempt_addtional_input_count > 0) {
-    const next_sub_input_items = [ ...sub_input_items, ...input_items.slice(offset, offset + next_attempt_addtional_input_count) ];
-    const added_count = Math.min(input_items.length - offset, next_attempt_addtional_input_count);
-    const done = input_items.length - offset <= next_attempt_addtional_input_count;
+    const next_sub_inputs = [ ...sub_inputs, ...inputs.slice(offset, offset + next_attempt_addtional_input_count) ];
+    const added_count = Math.min(inputs.length - offset, next_attempt_addtional_input_count);
+    const done = inputs.length - offset <= next_attempt_addtional_input_count;
     const payout_builder_context = {
       getOutputMinAmount: context.getOutputMinAmount.bind(context),
       getPreferredTokenOutputBCHAmount: context.getPreferredTokenOutputBCHAmount.bind(context),
       calcTxFeeWithOutputs: (outputs: Output[]): bigint => {
         const result = libauth.generateTransaction({
-          locktime: 0,
-          version: 2,
-          inputs: next_sub_input_items.map((a) => a.input), outputs: makeLAOutputs(outputs),
+          locktime: txparams.locktime,
+          version: txparams.version,
+          inputs: next_sub_inputs.map((a, i) => inputParamsWithUnlockerToLibauthInputTemplate(i, a, next_sub_inputs, outputs, txparams)),
+          outputs: outputs.map(outputToLibauthOutput),
         });
         if (!result.success) {
           /* c8 ignore next */
@@ -173,7 +181,7 @@ const createPayoutChainedTxSub = (context: CreatePayoutTxContext, input_items: L
       },
     };
     const sub_payout_rules = done ? payout_rules : [change_payout_rule];
-    const available_payouts: Array<{ token_id: TokenId, amount: bigint }> = calcAvailablePayoutFromLASourceOutputsAndOutputs(next_sub_input_items.map((a) => a.source_output), []);
+    const available_payouts: Array<{ token_id: TokenId, amount: bigint }> = calcAvailablePayoutFromIO(next_sub_inputs, []);
     if (available_payouts.filter((a) => a.amount < 0n).length > 0) {
       throw new InvalidProgramState(`Sum of the inputs & outputs is negative for the following token(s): ${available_payouts.filter((a) => a.amount < 0n).map((a) => a.token_id).join(', ')}`);
     }
@@ -181,10 +189,12 @@ const createPayoutChainedTxSub = (context: CreatePayoutTxContext, input_items: L
     if (token_burns.length > 0) {
       throw new ValueError(`Token burns are not allowed.`);
     }
+    const outputs = payout_outputs.map((a) => a.output);
     const result = libauth.generateTransaction({
-      locktime: 0,
-      version: 2,
-      inputs: next_sub_input_items.map((a) => a.input), outputs: makeLAOutputs(payout_outputs.map((a) => a.output)),
+      locktime: txparams.locktime,
+      version: txparams.version,
+      inputs: next_sub_inputs.map((a, i) => inputParamsWithUnlockerToLibauthInputTemplate(i, a, next_sub_inputs, outputs, txparams)),
+      outputs: outputs.map(outputToLibauthOutput),
     });
     if (!result.success) {
       /* c8 ignore next */
@@ -193,13 +203,13 @@ const createPayoutChainedTxSub = (context: CreatePayoutTxContext, input_items: L
     const txbin = libauth.encodeTransaction(result.transaction);
     if (txbin.length <= max_tx_size) {
       const txhash = libauth.hashTransactionUiOrder(txbin);
-      sub_input_items = next_sub_input_items;
-      unused_input_items = input_items.slice(offset + next_attempt_addtional_input_count);
+      sub_inputs = next_sub_inputs;
+      unused_inputs = inputs.slice(offset + next_attempt_addtional_input_count);
       best_result = {
         txbin, txhash, txfee,
         payouts: payout_outputs.map((a, i) => utxoFromPayoutResult(txhash, { output: a.output, output_index: i })),
         libauth_transaction: result.transaction,
-        libauth_source_outputs: sub_input_items.map((a) => a.source_output),
+        libauth_source_outputs: sub_inputs.map((a) => outputToLibauthOutput(a.utxo.output)),
       };
       best_result_done = done;
       if (done) {
@@ -213,5 +223,5 @@ const createPayoutChainedTxSub = (context: CreatePayoutTxContext, input_items: L
   if (best_result == null) {
     throw new InvalidProgramState('best_result == null');
   }
-  return { tx_result: best_result, done: best_result_done, unused_input_items };
+  return { tx_result: best_result, done: best_result_done, unused_inputs };
 }